// app/src/api/table/TableClient.ts
export interface Character {
  id: string;
  author: string;
  name: string;
  class_: string;
  strength: number;
  dexterity: number;
  constitution: number;
  intelligence: number;
  wisdom: number;
  charisma: number;
  hp_current: number;
  hp_max: number;
  created_at: number;
  updated_at: number;
}